extracted the anyme style  out of inputs
