+++ src/components/pages/profile/SVGstyle.js
@@ -1,0 +1,37 @@
+import React from "react";
+const SVGstyle = () =>{
+      return (
+        <svg className="svg" viewBox="0 0 320 550.8">
+        <defs>
+          <linearGradient
+            inkscapeCollect="always"
+            id="linearGradient"
+            x1="13"
+            y1="193.49992"
+            x2="307"
+            y2="193.49992"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop
+              style={{ stopColor: "#ff00ff" }}
+              offset="0"
+              id="stop876"
+            />
+            <stop
+              style={{ stopColor: "#ff0000" }}
+              offset="1"
+              id="stop878"
+            />
+          </linearGradient>
+        </defs>
+        <path
+          className="path"
+          d="M282.8,23l-240,0c0,0-25,0.8-25,35c0,34.2,25,35,25,35l-5.5,0l240,0c0,0,25,0.8,25,35c0,34.2-25,35-25,35l5.5,0l-240,0
+        c0,0-25,0.8-25,35c0,34.2,25,35,25,35l-5.5,0l240,0c0,0,25,0.8,25,35c0,34.2-25,35-25,35l8.3,0l-240,0c0,0-25,0.8-25,35
+        c0,34.2,25,35,25,35l-5.5,0l240,0c0,0,25,0.8,25,35c0,34.2-25,35-25,35H40c0,0-25,4-25,38.5S40,520,40,520h215c0,0,20-1,20-25
+        s-20-25-20-25H65c0,0-20,1.7-20,25c0,24,20,25,20,25h168.6"
+        />
+      </svg>
+      )
+}
+export default SVGstyle
